feat#verify3

+++ src/apis/cafeImage/cafeImage.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CafeImage } from './entities/cafeImage.entity';
@@ -19,7 +19,16 @@ export class CafeImageService {
     return result;
   }
 
-  async delete({ cafeImageID }) {
+  async delete({ cafeImageID, context }) {
+    const resultCafe = await this.cafeImageRepository.findOne({
+      where: {
+        id: cafeImageID,
+      },
+      relations: ['cafeInform', 'cafeInform.owner'],
+    });
+    if (resultCafe.cafeInform.owner.id !== context.req.user.id) {
+      throw new ConflictException('삭제 권한이 없습니다.');
+    }
     const result = await this.cafeImageRepository.delete({ id: cafeImageID });
     return result.affected ? true : false;
   }
